# apps/api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  notifications.BulkSendRequest:
    properties:
      channel:
        enum:
        - zalo_manual
        - zalo_zns
        - sms
        - zalo_personal
        type: string
      purpose:
        enum:
        - statement
        - statements
        - reminder
        type: string
    required:
    - purpose
    type: object
  notifications.BulkSendResponse:
    properties:
      bulk_text:
        description: |-
          BulkText joins every copy-paste row's message, separated by a blank
          line, so a teacher can copy once when pasting into a broadcast tool
          instead of copying each contact's message individually. Rows a
          zalo_personal run delivers itself are excluded.
        type: string
      collapsed_count:
        type: integer
      fallback_manual_count:
        type: integer
      personal_queued_count:
        description: |-
          PersonalQueuedCount and FallbackManualCount split a zalo_personal
          send's rows: mapped contacts the run delivers itself vs unmapped
          contacts left in BulkText for the teacher to copy-paste. Both zero for
          other channels.
        type: integer
      queued_count:
        type: integer
      rows:
        items:
          $ref: '#/definitions/notifications.BulkSendRow'
        type: array
      run_id:
        description: |-
          RunID names the background run auto-sending this call's zalo_personal
          rows; poll GET .../notifications/run for its progress. Null when
          nothing is being auto-sent — the copy-paste channels, or a personal
          send whose every contact was unmapped.
        type: string
      skipped_paid_count:
        type: integer
    type: object
  notifications.BulkSendRow:
    properties:
      channel:
        type: string
      collapsed:
        type: boolean
      contact_id:
        type: string
      contact_name:
        type: string
      message_text:
        type: string
      notification_id:
        type: string
      phone:
        type: string
      purpose:
        type: string
      status:
        type: string
      url:
        type: string
    type: object
  notifications.MarkSentRequest:
    properties:
      ids:
        items:
          type: string
        minItems: 1
        type: array
    required:
    - ids
    type: object
  notifications.NotificationResponse:
    properties:
      channel:
        type: string
      contact_id:
        type: string
      contact_name:
        type: string
      created_at:
        type: string
      error_message:
        description: |-
          ErrorMessage is a failed row's teacher-facing reason — the ledger is
          the only place a teacher can learn why a row was not delivered.
        type: string
      id:
        type: string
      phone:
        type: string
      purpose:
        type: string
      run_id:
        description: |-
          RunID pins the row to the paced run that sent it, letting a client
          show a run's failures without adopting older runs' rows.
        type: string
      sent_at:
        type: string
      status:
        type: string
    type: object
  notifications.RunSnapshotResponse:
    properties:
      active:
        type: boolean
      failed:
        type: integer
      purpose:
        type: string
      run_id:
        type: string
      sent:
        type: integer
      status:
        type: string
      total:
        type: integer
    type: object
  notifications.SendPreviewContact:
    properties:
      contact_id:
        type: string
      contact_name:
        type: string
    type: object
  notifications.SendPreviewResponse:
    properties:
      auto_send:
        description: |-
          AutoSend: mapped contacts who are friends of the caller's Zalo account
          — these rows would queue into a paced background run.
        items:
          $ref: '#/definitions/notifications.SendPreviewContact'
        type: array
      mapped_not_friend:
        description: |-
          MappedNotFriend: mapped, but the caller's account is not their friend —
          the DM may not reach them; they still queue as personal rows.
        items:
          $ref: '#/definitions/notifications.SendPreviewContact'
        type: array
      max_run_size:
        description: |-
          MaxRunSize is the server's cap on one run's auto-send count, so the
          client can warn about an oversized send before submitting it.
        type: integer
      unmapped:
        description: |-
          Unmapped: no Zalo mapping — these rows would fall back to the manual
          copy-paste channel.
        items:
          $ref: '#/definitions/notifications.SendPreviewContact'
        type: array
    type: object
...
  response.Envelope:
    properties:
      data: {}
      error:
        $ref: '#/definitions/response.ErrorBody'
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        type: boolean
    type: object
  response.ErrorBody:
    properties:
      code:
        type: string
      details:
        description: |-
          Details carries structured error context Fields' map[string]string
          cannot express — e.g. billing's list of unconfirmed sessions blocking a
          period close. Only ErrWithDetails populates it; Err leaves it nil.
      fields:
        additionalProperties:
          type: string
        type: object
      message:
        type: string
    type: object
  response.Meta:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
info:
  contact: {}
  description: Backend API for Teka. Every response uses the {success, data, meta,
    error} envelope.
  title: Teka API
  version: "1.0"
paths:
...
  /billing-periods/{id}/notifications:
    get:
      parameters:
      - description: billing period id
        in: path
        name: id
        required: true
        type: string
      - description: statements or reminder
        in: query
        name: purpose
        type: string
      - description: queued, sent, delivered, or failed
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/notifications.NotificationResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                error:
                  $ref: '#/definitions/response.ErrorBody'
              type: object
      security:
      - BearerAuth: []
      summary: List a billing period's notifications
      tags:
      - notifications
  /billing-periods/{id}/notifications/bulk:
    post:
      consumes:
      - application/json
      description: 'Regenerates/refreshes the period''s statements, then queues one
        message per eligible contact: purpose=statement (or "statements") targets
        every contact with a non-void invoice; purpose=reminder further narrows to
        contacts with outstanding > 0. One row per contact, never one per child.'
      parameters:
      - description: billing period id
        in: path
        name: id
        required: true
        type: string
      - description: purpose and optional channel
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/notifications.BulkSendRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/notifications.BulkSendResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                error:
                  $ref: '#/definitions/response.ErrorBody'
              type: object
        "403":
          description: caller lacks the send-reports permission
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                error:
                  $ref: '#/definitions/response.ErrorBody'
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                error:
                  $ref: '#/definitions/response.ErrorBody'
              type: object
        "409":
          description: period is not closed
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                error:
                  $ref: '#/definitions/response.ErrorBody'
              type: object
        "422":
          description: validation failed
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                error:
                  $ref: '#/definitions/response.ErrorBody'
              type: object
      security:
      - BearerAuth: []
      summary: Bulk-send statement notifications
      tags:
      - notifications
  /billing-periods/{id}/notifications/preview:
    get:
      description: 'Computes, without writing anything, the exact buckets a zalo_personal
        bulk send would produce right now: auto_send (mapped + friend of the caller''s
        Zalo account), mapped_not_friend, and unmapped (manual fallback), plus the
        server''s max_run_size cap. purpose=reminder narrows to contacts with outstanding
        > 0.'
      parameters:
      - description: billing period id
        in: path
        name: id
        required: true
        type: string
      - description: statements (default) or reminder
        in: query
        name: purpose
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/notifications.SendPreviewResponse'
              type: object
        "400":
          description: no linked Zalo account
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                error:
                  $ref: '#/definitions/response.ErrorBody'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                error:
                  $ref: '#/definitions/response.ErrorBody'
              type: object
        "403":
          description: caller lacks the send-reports permission
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                error:
                  $ref: '#/definitions/response.ErrorBody'
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                error:
                  $ref: '#/definitions/response.ErrorBody'
              type: object
        "409":
          description: cross-teacher period without the permission, or the Zalo session
            expired
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                error:
                  $ref: '#/definitions/response.ErrorBody'
              type: object
      security:
      - BearerAuth: []
      summary: Preview a zalo_personal bulk send
      tags:
      - notifications
  /billing-periods/{id}/notifications/run:
    get:
      description: Returns the period's latest background send run with progress counters
        derived from its rows. A period that never had a run answers active=false
        with a null run_id, not a 404 — poll this after a zalo_personal bulk send
        until active turns false.
      parameters:
      - description: billing period id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/notifications.RunSnapshotResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                error:
                  $ref: '#/definitions/response.ErrorBody'
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                error:
                  $ref: '#/definitions/response.ErrorBody'
              type: object
      security:
      - BearerAuth: []
      summary: Poll a period's zalo_personal run
      tags:
      - notifications
  /billing-periods/{id}/notifications/run/resume:
    post:
      description: Re-renders and re-sends only the run's still-queued rows; rows
        already sent keep their verdict. Only a run in the interrupted state qualifies.
        Rows that can no longer be auto-sent (mapping removed, balance since paid)
        are failed with a reason.
      parameters:
      - description: billing period id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/notifications.RunSnapshotResponse'
              type: object
        "400":
          description: no linked Zalo account
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                error:
                  $ref: '#/definitions/response.ErrorBody'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                error:
                  $ref: '#/definitions/response.ErrorBody'
              type: object
        "403":
          description: caller lacks the send-reports permission
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                error:
                  $ref: '#/definitions/response.ErrorBody'
              type: object
        "404":
          description: period has no run
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                error:
                  $ref: '#/definitions/response.ErrorBody'
              type: object
        "409":
          description: run is not interrupted, another run is sending, or the Zalo
            session expired
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                error:
                  $ref: '#/definitions/response.ErrorBody'
              type: object
      security:
      - BearerAuth: []
      summary: Resume an interrupted zalo_personal run
      tags:
      - notifications
...
  /notifications/mark-sent:
    post:
      consumes:
      - application/json
      description: 'Idempotent: an id already sent, or not belonging to the caller,
        is silently skipped rather than erroring.'
      parameters:
      - description: notification ids
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/notifications.MarkSentRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                error:
                  $ref: '#/definitions/response.ErrorBody'
              type: object
        "422":
          description: validation failed
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                error:
                  $ref: '#/definitions/response.ErrorBody'
              type: object
      security:
      - BearerAuth: []
      summary: Mark notifications sent
      tags:
      - notifications
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the access token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"